[Fix] 로그인 여부확인  헤더에서 Root로 변경
